requirements.txt
langchain
langchain-chroma
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
langgraph
langgraph-checkpoint
python-dotenv
numpy
pypdf
cryptography
duckduckgo_search
streamlit
chromadb


pysqlite3-binary